feat(ai-router): Schritt 2.4 Token-Budget-Manager
Neues Modul ai_router/budget.rs:
- BudgetManager mit per-UTC-Day und per-Session Tracking, Auto-Rollover
  beim Tageswechsel.
- BudgetLimits (daily_cents + optional session_cents), BudgetState
  Ok/Warn/Block (Warn ≥ 80%, Block ≥ 100%).
- cost_cents_for(Model) Tabelle aus docs/LLM-STRATEGIE.md — 0 für
  Abo-Modelle, ~1-13¢ für Pay-per-Use (fal.ai, Replicate, Kling-Pakete).
- UsageEntry history mit CSV-Export (timestamp, provider, model, task,
  cost_cents).
- would_block(projected_cents) für pre-call Gate; zero limits als
  "no limit" interpretiert.

Router-Integration (ai_router/mod.rs):
- AiRouter hält Arc<BudgetManager>. route() blockt mit
  RouterError::BudgetExceeded bevor jede Dispatches projected cost.
- Nach Success: record(UsageEntry) mit response.cost_cents oder
  Fallback-Tabelle.
- AiResponse.cost_cents auf u64 (war u32) — konsistent mit Budget.
- RouterError::BudgetExceeded + IPC-Mirror.

Tauri-Commands (alle registriert in lib.rs):
- get_budget_status → BudgetStatus
- set_budget_limit(limits) → aktualisierter Status
- export_usage → CSV-String

Frontend:
- src/lib/budgetCommands.ts typed invoke-Wrapper
- src/hooks/useBudgetPoll.ts pollt get_budget_status alle 5s, schreibt
  in aiStore.budget → Footer-StatusBar zeigt live $X.XX / $Y.YY.
- App.tsx mountet useBudgetPoll() neben den anderen Boot-Hooks.

15 neue Budget-Tests + 2 Router-Integration-Tests (BudgetExceeded
block, record-on-success) = 17 neu. 154 Rust-Tests total, 184 Front-
end unverändert. Clippy -D warnings clean, fmt clean.

+++ src-tauri/src/ai_router/models.rs
@@ -144,7 +144,7 @@ pub struct AiResponse {
     pub output: serde_json::Value,
     /// Cost in cents (integer) when known.
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub cost_cents: Option<u32>,
+    pub cost_cents: Option<u64>,
     /// True when served from the router's cache (Schritt 2.3).
     #[serde(default)]
     pub cached: bool,
